requirements.txt

fastapi == 0.111.0
# fastapi_jwt_auth == 0.5.0
uvicorn == 0.30.1
sqlalchemy == 2.0.30
# sqlalchemy-utils == 0.41.2
# mysql-connector-python == 8.4.0
mysqlclient == 2.2.4
# werkzeug == 3.0.3
pydantic-settings == 2.3.1
passlib[bcrypt]



